Move check file exist to util
